pyproject.toml
[project]
name = "agentic-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.4",
    "google-generativeai>=0.8.6",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
]
